fix: address keeper credential review comments
